+++ tests/Jroc.Tests/IntrinsicCallables/JavaScript/IntrinsicCallables_RegExp_Test_GuardedOverrides.js
@@ -1,0 +1,9 @@
+"use strict";
+
+const ownOverride = /a/;
+console.log(ownOverride.test("a"));
+ownOverride.test = () => "own";
+console.log(ownOverride.test("a"));
+
+RegExp.prototype.test = () => "prototype";
+console.log(/a/.test("a"));